+++ src/parkapi_sources/converters/bfrk_bw/base_converter.py
@@ -22,11 +22,17 @@ class BfrkBasePushConverter(PullConverter, ABC):
     def bfrk_validator(self) -> DataclassValidator:
         pass
 
+    @property
+    @abstractmethod
+    def source_url_config_key(self) -> str:
+        pass
+
     def get_static_parking_sites(self) -> tuple[list[StaticParkingSiteInput], list[ImportParkingSiteException]]:
         static_parking_site_inputs: list[StaticParkingSiteInput] = []
         static_parking_site_errors: list[ImportParkingSiteException] = []
 
-        response = requests.get(self.source_info.source_url, timeout=300)
+        source_url = self.config_helper.get(self.source_url_config_key, self.source_info.source_url)
+        response = requests.get(source_url, timeout=300)
 
         input_dicts = response.json()